+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 ## v0.0.9
 
-* Add tests.
+* Implement tests for DocFXProject.
+
+## v0.1.0
+
+* Implement MetadataCache.
+* Implement tests for MetadataCache.